Workaround for vkGetPhysicalDeviceToolProperties
VK_EXT_tooling_info is a device extension, so we must load
vkGetPhysicalDeviceToolPropertiesEXT from the device.  However, when
promoted to core Vulkan 1.3, the command must be loaded from the
instance.  So as to not break the design goal of command aliases being
equivalent, work around this by loading the command from the instance,
but store the command on the device.

This lets client code call get_physical_device_tool_properties on the
device regardless of whether the command is there due to core version
being >= 1.3 or VK_EXT_tooling_info.